Refactor getter
